pyproject.toml
[tool.poetry]
name = "wrs-graduation-research"
version = "0.1.0"
description = ""
authors = ["weiwei wan","Haruto Nagai"]
# vital
packages = [
    { include =  "wrs", from =  "." },
    { include =  "nagai", from =  "." }
]
[tool.poetry.dependencies]
# open3d does not work on Python 3.11+
python = ">=3.9,<3.12"
numpy = "^1.26.1"
# not 1.10.13.post but 1.10.13
panda3d = "1.10.13"
opencv-python = "^4.8.1.78"
opencv-contrib-python = "^4.8.1.78"
scikit-learn = "^1.3.2"
grpcio = "^1.59.2"
grpcio-tools = "^1.59.2"
pyyaml = "^6.0.1"
rtree = "^1.1.0"
open3d = "^0.18.0"
shapely = "^2.0.2"
networkx = "^3.2.1"
pyserial = "^3.5"
pandas = "^2.1.2"
matplotlib = "^3.8.0"
cython = "^3.0.5"
gimpact = "^1.0.1"
# The dependencies of open3d
addict = "^2.4.0"
tqdm = "^4.66.1"
isort = "^5.12.0"
black = "^23.11.0"
flake8 = "^6.1.0"
pddlpy = "^0.3.4"
pddl = "^0.4.0"
deepdiff = "^6.7.1"
autoflake = "^2.2.1"
stl = "^0.0.3"
numpy-stl = "^3.1.1"
pygame = "^2.5.2"
scikit-image = "^0.22.0"
ikpy = "^3.3.4"
keyboard = "^0.13.5"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"